+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fixed
+- Fixed required Payment fields being validated and processed on intermediate multi-page form submissions. ([#2873](https://github.com/verbb/formie/issues/2873))
 - Fixed a validation error being raised for conditionally-hidden sub-fields in a Repeater when the same field was visible in another row. ([#2874](https://github.com/verbb/formie/issues/2874))
 - Fixed conditionally-hidden fields inside a Repeater not being revealed when they have a validation error on Ajax forms with a Stripe payment field. ([#2875](https://github.com/verbb/formie/issues/2875))
 - Fixed missing authorization for `integrations/form-settings`, which could allow SSRF and exposure of integration credentials.